Animal-Kingdom/Animal_Kingdom/pose_estimation/code/requirements.txt
absl-py==2.1.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.9
dataclasses==0.6
easydict==1.13
einops==0.7.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
future==1.0.0
gdown==5.1.0
google-api-core==2.17.1
google-auth==2.28.1
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-core==2.4.1
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.2
importlib-resources==6.1.3
Jinja2==3.1.3
json-tricks==3.17.3
kiwisolver==1.4.5
lazy-loader==0.3
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.19.3
# nvidia-nvjitlink-cu12==12.4.99
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pillow==10.2.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
# pycocotools==2.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.3
six==1.16.0
soupsieve==2.5
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tifffile==2023.7.10
# # timm==0.4.9
# torch==1.8.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
# torchvision==0.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html
# torchaudio==0.8.0 -f https://download.pytorch.org/whl/torch_stable.html
tqdm==4.66.2
# triton
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
yacs==0.1.8
zipp==3.17.0
